revise dir
