+++ tatk/policy/mle/multiwoz/mle.py
@@ -4,34 +4,26 @@ import os
 import json
 from tatk.policy.policy import Policy
 from tatk.policy.rlmodule import MultiDiscretePolicy
-from tatk.policy.multiwoz.vector_multiwoz import MultiWozVector
-from tatk.policy.camrest.vector_camrest import CamrestVector
+from tatk.policy.vector.vector_multiwoz import MultiWozVector
 
 DEVICE = torch.device("cuda" if torch.cuda.is_available() else "cpu")
 
 class MLP(Policy):
     
-    def __init__(self, is_train=False, dataset='multiwoz'):
+    def __init__(self, is_train=False):
         root_dir = os.path.dirname(os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__)))))
         self.is_train = is_train
         
         with open(os.path.join(os.path.dirname(os.path.abspath(__file__)), 'config.json'), 'r') as f:
-            cfg = json.load(f)[dataset]
+            cfg = json.load(f)
         
-        if dataset == 'multiwoz':
-            voc_file = os.path.join(root_dir, 'data/multiwoz/sys_da_voc.txt')
-            voc_opp_file = os.path.join(root_dir, 'data/multiwoz/usr_da_voc.txt')
-            self.vector = MultiWozVector(voc_file, voc_opp_file)
-        elif dataset == 'camrest':
-            voc_file = os.path.join(root_dir, 'data/camrest/sys_da_voc.txt')
-            voc_opp_file = os.path.join(root_dir, 'data/camrest/usr_da_voc.txt')
-            self.vector = CamrestVector(voc_file, voc_opp_file)
-        else:
-            raise NotImplementedError('unknown dataset {}'.format(dataset))
+        voc_file = os.path.join(root_dir, 'data/multiwoz/sys_da_voc.txt')
+        voc_opp_file = os.path.join(root_dir, 'data/multiwoz/usr_da_voc.txt')
+        self.vector = MultiWozVector(voc_file, voc_opp_file)
                
         self.policy = MultiDiscretePolicy(self.vector.state_dim, cfg['h_dim'], self.vector.da_dim).to(device=DEVICE)
         
-    def predict(self, state, sess=None):
+    def predict(self, state):
         """
         Predict an system action given state.
         Args:
