+++ tatk/policy/ppo/ppo.py
@@ -21,7 +21,7 @@ class PPO(Policy):
         self.policy = MultiDiscretePolicy(self.vector.state_dim, cfg['h_dim'], self.vector.da_dim).to(device=DEVICE)
         self.value = Value(self.vector.state_dim, cfg['hv_dim']).to(device=DEVICE)
         
-    def predict(self, state, sess=None):
+    def predict(self, state):
         """
         Predict an system action given state.
         Args:
